+++ .github/workflows/review-build-image.yml
@@ -9,7 +9,7 @@
 # No terraform/jd state is read, so the publish role needs only ECR push.
 #
 # Prerequisites (one-time, by a maintainer):
-#   - Deploy the tf-aws-review-ci template (`jd init`/`jd up`), which creates the
+#   - Deploy the tf-aws-iam-review template (`jd init`/`jd up`), which creates the
 #     ECR repo and the publish role.
 #   - Set these repo variables from the template outputs. Until both are set,
 #     this workflow is inert:
