+++ libs/jupyter-infra-tf-aws-iam-review/jupyter_infra_tf_aws_iam_review/template/AGENT.md.template
@@ -1,4 +1,4 @@
-# Jupyter-deploy: Review CI Infrastructure Template (tf-aws-review-ci)
+# Jupyter-deploy: Review CI Infrastructure Template (tf-aws-iam-review)
 
 ## Project organization
