+++ libs/jupyter-infra-tf-aws-iam-review/jupyter_infra_tf_aws_iam_review/template/engine/modules/ecr_repository/variables.tf
